renamed bone_mineral_density module to bone_parameters
