+++ ormir_xct/bone_parameters/README.md
@@ -1,3 +1,3 @@
-# bone_mineral_density
+# bone_parameters
 Implementation of the IPL BMD commands (/vox_gobj_scanco_param and 
 /vox_scanco_param).
